restore cache in place
Copying the artifact to a temporary and isolated /nix/store isn't
necessary if we are using exclusively the cache. All we need to do is
ensure that the key restored on the runners is an exact cache hit,
which should always be true since we just created it right before
they spawn.