fix caching of build time dependencies
462c4ca7974448b6b7fc6011db4b452afb25af55 accidentally broke caching of
uncached build time dependencies. However, we still don't want to
invoke `nix-build` more than once for performance reasons.

The solution is to use the previous method for the caching step
combined with `xargs`, which ensures the OS arg limit is never exceeded
by default, while maintain the new method for the build step.