run: fix a bad assumption for actions
In order to attempt to ensure that a nix evaluation never occurs on
a job runner, an assumption was made that action derivations would
always simply invoke their target derivations. This is not actually
the case so we have to just build and call the `actionDrv` to ensure
the proper action is run.

Unfortunately this means that the action may call Nix and may induce
evaluation, but in practice no jobs seem to introduce a significant
eval overhead as of now.

However, we may want to explore a solution upstream where we track
any derivations the action may invoke, so we have a triplet of:
- actionDrv
- targetDrv
- actionTargetDrv

Which would then allows us to ensure we don't introduce another,
potentially expensive, evaluation.