fix: drop target dedupliaction
Target deduplication is a view-port mutation that might deteriorate
the trust of users since they don't see something anywhere that they
would reasonably expect to see.

While that makes sense from a deduplication standpoint in the CI, we can't control which one is going to be removed (foo or default) and for the end user not seeing a target (for otherwise unknown reasons) might be unnerving and diminish trust in std-action.