fix!: mark jobs as skipped instead of removing them
This allows a situation where a job is required for a PR merge to still
pass if they are skipped as per the documentation here:
https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/collaborating-on-repositories-with-code-quality-features/troubleshooting-required-status-checks#handling-skipped-but-required-checks

BREAKING CHANGE: a new job conditional will be required
```
if: matrix.target.skipped == true
```
Otherwise the job will still run. This allows GitHub to properly update
the status of the job without actually having to run it.